# api/openapi.yaml
openapi: 3.0.3
info:
  title: Journey Platform API
  description: Event-Driven Journey Engine & Workflow Platform REST API Specification
  version: 1.0.0
servers:
  - url: /api/v1
    description: Base v1 API path
paths:
...
  /catalogs/events:
    get:
      summary: List event catalog records
      operationId: listEventCatalog
      tags:
        - Catalogs
      parameters:
        - $ref: '#/components/parameters/HeaderRequestID'
      responses:
        '200':
          description: List of event catalog records
          headers:
            Request-ID:
              $ref: '#/components/headers/HeaderRequestIDResponse'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogRecord'
        '400':
          $ref: '#/components/responses/ErrorResponse400'
        '500':
          $ref: '#/components/responses/ErrorResponse500'
  /catalogs/actions:
    get:
      summary: List action catalog records
      operationId: listActionCatalog
      tags:
        - Catalogs
      parameters:
        - $ref: '#/components/parameters/HeaderRequestID'
      responses:
        '200':
          description: List of action catalog records
          headers:
            Request-ID:
              $ref: '#/components/headers/HeaderRequestIDResponse'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogRecord'
        '400':
          $ref: '#/components/responses/ErrorResponse400'
        '500':
          $ref: '#/components/responses/ErrorResponse500'
  /catalogs/attributes:
    get:
      summary: List attribute catalog records
      operationId: listAttributeCatalog
      tags:
        - Catalogs
      parameters:
        - $ref: '#/components/parameters/HeaderRequestID'
      responses:
        '200':
          description: List of attribute catalog records
          headers:
            Request-ID:
              $ref: '#/components/headers/HeaderRequestIDResponse'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogRecord'
        '400':
          $ref: '#/components/responses/ErrorResponse400'
        '500':
          $ref: '#/components/responses/ErrorResponse500'
  /catalogs/parameters:
    get:
      summary: List parameter catalog records
      operationId: listParameterCatalog
      tags:
        - Catalogs
      parameters:
        - $ref: '#/components/parameters/HeaderRequestID'
      responses:
        '200':
          description: List of parameter catalog records
          headers:
            Request-ID:
              $ref: '#/components/headers/HeaderRequestIDResponse'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogRecord'
        '400':
          $ref: '#/components/responses/ErrorResponse400'
        '500':
          $ref: '#/components/responses/ErrorResponse500'
...
components:
  parameters:
    HeaderRequestID:
      name: Request-ID
      in: header
      required: false
      description: Unique request identifier for distributed tracing
      schema:
        type: string
        example: "req-12345678"
...
  headers:
    HeaderRequestIDResponse:
      description: Unique request identifier returned by the server
      schema:
        type: string
        example: "req-12345678"
...
  responses:
    ErrorResponse400:
      description: Invalid client request body or parameters
      headers:
        Request-ID:
          $ref: '#/components/headers/HeaderRequestIDResponse'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ErrorResponse500:
      description: Internal server error
      headers:
        Request-ID:
          $ref: '#/components/headers/HeaderRequestIDResponse'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CatalogRecord:
      type: object
      required:
        - schema_version
        - record_id
        - name
        - component_type
        - version
        - description
        - tags
        - is_deprecated
      properties:
        schema_version:
          type: string
          enum: ["1.0"]
        record_id:
          type: string
        name:
          type: string
        component_type:
          type: string
          enum: [activity, trigger, action, condition]
        version:
          type: string
        description:
          type: string
        schema_definition:
          type: object
        content_hash:
          type: string
        tags:
          type: array
          items:
            type: string
        is_deprecated:
          type: boolean
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          example: "INVALID_REQUEST"
        message:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
        request_id:
          type: string
          example: "req-12345678"
      additionalProperties: false